simple user ui 구현
