+++ android/app/src/main/java/com/example/stacklounge/AdaptermainFragment.kt
@@ -3,10 +3,10 @@ package com.example.stacklounge
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentManager
 import androidx.fragment.app.FragmentStatePagerAdapter
-import com.example.stacklounge.Fragment.FragmentMainCommunity
-import com.example.stacklounge.Fragment.FragmentMainFavorite
-import com.example.stacklounge.Fragment.FragmentMainSearch
-import com.example.stacklounge.Fragment.FragmentMainUser
+import com.example.stacklounge.fragment.FragmentMainCommunity
+import com.example.stacklounge.fragment.FragmentMainFavorite
+import com.example.stacklounge.fragment.FragmentMainSearch
+import com.example.stacklounge.fragment.FragmentMainUser
 
 class AdapterMainFragment(fm : FragmentManager, private val fragmentCount : Int) : FragmentStatePagerAdapter(fm) {
     override fun getItem(position: Int) : Fragment {
