+++ android/app/src/main/java/com/example/stacklounge/fragment/FragmentMainFavorite.kt
@@ -1,4 +1,4 @@
-package com.example.stacklounge.Fragment
+package com.example.stacklounge.fragment
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -14,10 +14,10 @@ private const val ARG_PARAM2 = "param2"
 
 /**
  * A simple [Fragment] subclass.
- * Use the [FragmentMainUser.newInstance] factory method to
+ * Use the [FragmentMainFavorite.newInstance] factory method to
  * create an instance of this fragment.
  */
-class FragmentMainUser : Fragment() {
+class FragmentMainFavorite : Fragment() {
     // TODO: Rename and change types of parameters
     private var param1: String? = null
     private var param2: String? = null
@@ -35,7 +35,7 @@ class FragmentMainUser : Fragment() {
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_main_user, container, false)
+        return inflater.inflate(R.layout.fragment_main_favorite, container, false)
     }
 
     companion object {
@@ -45,12 +45,12 @@ class FragmentMainUser : Fragment() {
          *
          * @param param1 Parameter 1.
          * @param param2 Parameter 2.
-         * @return A new instance of fragment FragmentMainUser.
+         * @return A new instance of fragment FragmentMainFavorite.
          */
         // TODO: Rename and change types and number of parameters
         @JvmStatic
         fun newInstance(param1: String, param2: String) =
-            FragmentMainUser().apply {
+            FragmentMainFavorite().apply {
                 arguments = Bundle().apply {
                     putString(ARG_PARAM1, param1)
                     putString(ARG_PARAM2, param2)
