+++ android/app/src/main/java/com/example/stacklounge/fragment/FragmentMainUser.kt
@@ -1,11 +1,18 @@
-package com.example.stacklounge.Fragment
+package com.example.stacklounge.fragment
 
+import android.content.Intent
 import android.os.Bundle
-import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import android.widget.Button
+import androidx.fragment.app.Fragment
 import com.example.stacklounge.R
+import com.example.stacklounge.login.LoginActivity
+import com.google.firebase.auth.ktx.auth
+import com.google.firebase.ktx.Firebase
+import kotlinx.android.synthetic.main.fragment_main_user.*
+
 
 // TODO: Rename parameter arguments, choose names that match
 // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
@@ -14,10 +21,10 @@ private const val ARG_PARAM2 = "param2"
 
 /**
  * A simple [Fragment] subclass.
- * Use the [FragmentMainFavorite.newInstance] factory method to
+ * Use the [FragmentMainUser.newInstance] factory method to
  * create an instance of this fragment.
  */
-class FragmentMainFavorite : Fragment() {
+class FragmentMainUser : Fragment() {
     // TODO: Rename and change types of parameters
     private var param1: String? = null
     private var param2: String? = null
@@ -28,6 +35,7 @@ class FragmentMainFavorite : Fragment() {
             param1 = it.getString(ARG_PARAM1)
             param2 = it.getString(ARG_PARAM2)
         }
+
     }
 
     override fun onCreateView(
@@ -35,7 +43,17 @@ class FragmentMainFavorite : Fragment() {
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_main_favorite, container, false)
+
+        /*btnlogout.setOnClickListener {
+            Firebase.auth.signOut()
+
+            startActivity(
+                Intent(getActivity(), LoginActivity::class.java)
+            )
+
+        }*/
+
+        return inflater.inflate(R.layout.fragment_main_user, container, false)
     }
 
     companion object {
@@ -45,12 +63,12 @@ class FragmentMainFavorite : Fragment() {
          *
          * @param param1 Parameter 1.
          * @param param2 Parameter 2.
-         * @return A new instance of fragment FragmentMainFavorite.
+         * @return A new instance of fragment FragmentMainUser.
          */
         // TODO: Rename and change types and number of parameters
         @JvmStatic
         fun newInstance(param1: String, param2: String) =
-            FragmentMainFavorite().apply {
+            FragmentMainUser().apply {
                 arguments = Bundle().apply {
                     putString(ARG_PARAM1, param1)
                     putString(ARG_PARAM2, param2)
