+++ android/app/src/main/res/layout/fragment_main_community.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Fragment.FragmentMainCommunity">
+    tools:context=".fragment.FragmentMainCommunity">
 
 
     <androidx.constraintlayout.widget.ConstraintLayout
