+++ android/app/src/main/res/layout/fragment_main_favorite.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Fragment.FragmentMainFavorite">
+    tools:context=".fragment.FragmentMainFavorite">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
