+++ android/app/src/main/res/layout/fragment_main_search.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Fragment.FragmentMainSearch">
+    tools:context=".fragment.FragmentMainSearch">
 
     <ScrollView
         android:layout_width="match_parent"
@@ -35,6 +35,7 @@
                android:layout_marginEnd="20dp"
                android:layout_marginBottom="8dp"
                android:layout_marginTop="8dp"
+               android:padding="20dp"
                android:background="@drawable/square_border_black"
                android:src="@drawable/baemin" />
 
@@ -46,6 +47,7 @@
                android:layout_marginEnd="20dp"
                android:layout_marginBottom="8dp"
                android:layout_marginTop="8dp"
+               android:padding="30dp"
                android:background="@drawable/square_border_black"
                android:src="@drawable/socar" />
